+++ DSP/pattern.cpp
@@ -3,9 +3,9 @@
 Pattern::Pattern(int ChannelCount)
 {
     this->ChannelCount = ChannelCount;
-    data = new QVector<double>*[ChannelCount];
+    data = new QVector<float>*[ChannelCount];
     for(int i=0;i<ChannelCount;++i)
-        data[i] = new QVector<double>();
+        data[i] = new QVector<float>();
 }
 
 Pattern::~Pattern()
@@ -17,12 +17,12 @@ Pattern::~Pattern()
     delete[] data;
 }
 
-void Pattern::Put(int channel, double val)
+void Pattern::Put(int channel, float val)
 {
     data[channel]->push_back(val);
 }
 
-double Pattern::Get(int channel, int pos)
+float Pattern::Get(int channel, int pos)
 {
     if(pos >=0 && pos < data[channel]->count())
         return data[channel]->at(pos);
