+++ DSP/splitter.cpp
@@ -9,7 +9,7 @@ void Splitter::Update()
 {
     for(int i=0;i<InputChannelCount;++i)
     {
-        double val = ReadInput(i);
+        float val = ReadInput(i);
         for(int j=i;j<OutputChannelCount;j+=InputChannelCount)
             WriteOutput(j,val);
     }
